requirements.txt
jinja2
python-rofi
pyyaml